requirements.txt
Package            Version
------------------ ---------
aiofiles           23.1.0
aiogram            3.1.1
aiohttp            3.8.6
aiosignal          1.3.1
annotated-types    0.6.0
async-timeout      4.0.3
attrs              23.1.0
certifi            2023.7.22
charset-normalizer 3.3.0
frozenlist         1.4.0
idna               3.4
magic-filter       1.0.12
multidict          6.0.4
pip                21.1.1
psycopg2-binary    2.9.9
pyaes              1.6.1
pydantic           2.3.0
pydantic-core      2.6.3
Pyrogram           2.0.106
PySocks            1.7.1
setuptools         56.0.0
typing-extensions  4.7.1
yarl               1.9.2
